Continue refactoring study page.
